requirements.txt
alembic==1.4.2
Babel==2.8.0
backcall==0.1.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
decorator==4.4.2
dominate==2.5.1
elasticsearch==7.6.0
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
guess-language-spirit==0.5.3
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pyenchant==3.0.1
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
redis==3.4.1
requests==2.23.0
rq==1.3.0
six==1.14.0
SQLAlchemy==1.3.15
traitlets==4.3.3
urllib3==1.25.8
visitor==0.1.3
wcwidth==0.1.9
Werkzeug==1.0.0
WTForms==2.2.1
